pyproject.toml
[project]
name = "asset-manager"
version = "0.1.2"
description = ""
authors = [
    {name = "TheUltimateChickenLord"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry]
packages = [{include = "asset_manager"}]

[tool.poetry.group.dev.dependencies]
conventional-commit = ">=0.4.2,<0.5.0"
pylint = "^3.3.7"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
